moved edit and delete buttons to show inside post boxes

+++ public/javascript/delete-comment.js
@@ -4,15 +4,17 @@ async function deleteComment(event) {
     const comment_id = event.target.getAttribute('data-id');
     console.log("click " + comment_id);
 
-    const response = await fetch(`/api/comments/${comment_id}`, {
-        method: 'DELETE'
-    });
+    if (comment_id) {
+        const response = await fetch(`/api/comments/${comment_id}`, {
+            method: 'DELETE'
+        });
 
-    if (response.ok) {
-        document.location.reload();
-    }
-    else {
+        if (response.ok) {
+         document.location.reload();
+        }
+        else {
         alert(response.statusText);
+        }
     }
 }
